Crate kimchi_optimism

Re-exports

pub use ramlookup::LookupMode as RAMLookupMode;

pub use ramlookup::RAMLookup;

Modules

cannon

Modules mimicking the defined structures used by Cannon CLI.

cannon_cli

A CLI mimicking the Cannon CLI.

keccak

Implementation of Keccak used by the zkVM.

lookups

Instantiation of the lookups for the VM project. Instantiation of the lookups for the VM project.

mips

MIPS interpreter. This module implements a zero-knowledge virtual machine (zkVM) for the MIPS architecture. A zkVM is used by a prover to convince a verifier that the execution trace (also called the witness) of a program execution is correct. In the case of this zkVM, we will represent the execution trace by using a set of columns whose values will represent the evaluations of polynomials over a certain pre-defined domain. The correct execution will be proven using a polynomial commitment protocol. The polynomials are described in the structure crate::mips::column::Column. These polynomials will be committed and evaluated at certain points following the polynomial protocol, and it will form the proof of the correct execution that the prover will build and send to the verifier. The corresponding structure is crate::proof::Proof. The prover will start by computing the execution trace using the interpreter implemented in the module crate::mips::interpreter, and the evaluations will be kept in the structure crate::proof::ProofInputs.

preimage_oracle

Preimage oracle interface used by the zkVM.

proof

Proof system of the zkVM.

ramlookup

The RAM lookup argument.

witness

Generic definition of a zkvm witness.

Constants

DOMAIN_SIZE

Domain size shared by the Keccak evaluations, MIPS evaluation and main program.