Module kimchi_optimism::keccak::column

This module defines the custom columns used in the Keccak witness, which are aliases for the actual Keccak witness columns also defined here.

Enums

Column

Column aliases used by the Keccak circuit. The number of aliases is not necessarily equal to the actual number of columns. Each alias will be mapped to a column index depending on the step kind (Sponge or Round) that is currently being executed.

Constants

ZKVM_KECCAK_COLS

The total number of witness columns used by the Keccak circuit.

Traits

KeccakWitnessTrait

Type Aliases

KeccakWitness

The witness columns used by the Keccak circuit. The Keccak circuit is split into two main modes: Sponge and Round. The columns are shared between the Sponge and Round steps. The hash and step indices are shared between both modes. The row is split into the following entries: