Trait kimchi_optimism::keccak::column::KeccakWitnessTrait

pub trait KeccakWitnessTrait<T> {
    // Required methods
    fn hash_index(&self) -> &T;
    fn block_index(&self) -> &T;
    fn step_index(&self) -> &T;
    fn mode_flags(&self) -> &[T];
    fn mode_flags_mut(&mut self) -> &mut [T];
    fn curr(&self) -> &[T];
    fn curr_mut(&mut self) -> &mut [T];
    fn next(&self) -> &[T];
    fn next_mut(&mut self) -> &mut [T];
    fn chunk(&self, offset: usize, length: usize) -> &[T];
}

Required Methods

fn hash_index(&self) -> &T

Returns the hash index

fn block_index(&self) -> &T

Returns the block index

fn step_index(&self) -> &T

Returns the step index

fn mode_flags(&self) -> &[T]

Returns the mode flags

fn mode_flags_mut(&mut self) -> &mut [T]

Returns the mode flags as a mutable reference

fn curr(&self) -> &[T]

Returns the curr witness columns

fn curr_mut(&mut self) -> &mut [T]

Returns the curr witness columns as a mutable reference

fn next(&self) -> &[T]

Returns the next witness columns

fn next_mut(&mut self) -> &mut [T]

Returns the next witness columns as a mutable reference

fn chunk(&self, offset: usize, length: usize) -> &[T]

Returns a chunk of the curr witness columns