Type Alias kimchi_optimism::keccak::column::KeccakWitness

pub type KeccakWitness<T> = Witness<ZKVM_KECCAK_COLS, T>;

The witness columns used by the Keccak circuit. The Keccak circuit is split into two main modes: Sponge and Round. The columns are shared between the Sponge and Round steps. The hash and step indices are shared between both modes. The row is split into the following entries:

hash_index: Which hash this is inside the circuit

step_index: Which step this is inside the hash

mode_flags: Round, Absorb, Squeeze

curr: Contains 1969 witnesses used in the current step including Input and RoundConstants

next: Contains the Output

Aliased Type

struct KeccakWitness<T> {
    pub cols: [T; 2074],
}

Fields
cols: [T; 2074]

A witness row is represented by an array of N witness columns When T is a vector, then the witness describes the rows of the circuit.

Implementations

impl<T: Clone> KeccakWitness<T>

pub fn step_index(&self) -> &T

pub fn chunk(&self, offset: usize, length: usize) -> &[T]

Returns a chunk of the curr witness columns