Trait kimchi_optimism::keccak::constraints::Constraints

pub trait Constraints {
    type Column;
    type Variable: Mul<Self::Variable, Output = Self::Variable> + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Clone;
    type Fp: Neg<Output = Self::Fp>;

    // Required methods
    fn variable(&self, column: Self::Column) -> Self::Variable;
    fn constrain(&mut self, x: Self::Variable);
    fn constraints(&mut self);
}

This trait contains the constraints for one Keccak step.

Required Associated Types

type Column

type Variable: Mul<Self::Variable, Output = Self::Variable> + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Clone

type Fp: Neg<Output = Self::Fp>

Required Methods

fn variable(&self, column: Self::Column) -> Self::Variable

Returns the variable corresponding to a given column alias.

fn constrain(&mut self, x: Self::Variable)

Adds one constraint to the environment.

fn constraints(&mut self)

Adds all 887 constraints to the environment and triggers read lookups:

143 constraints of degree 1

739 constraints of degree 2

5 constraints of degree 5

Object Safety

This trait is not object safe.