Trait kimchi_optimism::keccak::interpreter::KeccakInterpreter

pub trait KeccakInterpreter<F: One + Debug + Zero> {
    type Variable: Mul<Self::Variable, Output = Self::Variable> + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Clone + Debug + One + Zero;

    // Required methods
    fn constant(x: u64) -> Self::Variable;
    fn constant_field(x: F) -> Self::Variable;
    fn two_pow(x: u64) -> Self::Variable;

    // Provided methods
    fn is_boolean(x: Self::Variable) -> Self::Variable { ... }
    fn not(x: Self::Variable) -> Self::Variable { ... }
    fn is_one(x: Self::Variable) -> Self::Variable { ... }
    fn is_nonzero(x: Self::Variable, x_inv: Self::Variable) -> Self::Variable { ... }
    fn xor(x: Self::Variable, y: Self::Variable) -> Self::Variable { ... }
    fn or(x: Self::Variable, y: Self::Variable) -> Self::Variable { ... }
    fn either_zero(x: Self::Variable, y: Self::Variable) -> Self::Variable { ... }
    fn zero() -> Self::Variable { ... }
    fn one() -> Self::Variable { ... }
    fn two() -> Self::Variable { ... }

}

This trait includes functionalities needed to obtain the variables of the Keccak circuit needed for constraints and witness

Required Associated Types

type Variable: Mul<Self::Variable, Output = Self::Variable> + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Clone + Debug + One + Zero

Required Methods

fn constant(x: u64) -> Self::Variable

Creates a variable from a constant integer

fn constant_field(x: F) -> Self::Variable

Creates a variable from a constant field element

fn two_pow(x: u64) -> Self::Variable

Returns a variable representing the value 2^x

Provided Methods

fn is_boolean(x: Self::Variable) -> Self::Variable

Degree-2 variable encoding whether the input is a boolean value (0 = yes)

fn not(x: Self::Variable) -> Self::Variable

Degree-1 variable encoding the negation of the input Note: it only works as expected if the input is a boolean value

fn is_one(x: Self::Variable) -> Self::Variable

Degree-1 variable encoding whether the input is the value one (0 = yes)

fn is_nonzero(x: Self::Variable, x_inv: Self::Variable) -> Self::Variable

Degree-2 variable encoding whether the first input is nonzero (0 = yes). It requires the second input to be the multiplicative inverse of the first. Note: if the first input is zero, there is no multiplicative inverse.

fn xor(x: Self::Variable, y: Self::Variable) -> Self::Variable

Degree-1 variable encoding the XOR of two variables which should be boolean (1 = true)

fn or(x: Self::Variable, y: Self::Variable) -> Self::Variable

Degree-1 variable encoding the OR of two variables, which should be boolean (1 = true)

fn either_zero(x: Self::Variable, y: Self::Variable) -> Self::Variable

Degree-2 variable encoding whether at least one of the two inputs is zero (0 = yes)

fn zero() -> Self::Variable

Returns a variable representing the value zero

fn one() -> Self::Variable

Returns a variable representing the value one

fn two() -> Self::Variable

Returns a variable representing the value two

Object Safety

This trait is not object safe.