Struct kimchi_optimism::keccak::witness::Env

pub struct Env<F> {
    pub witness: KeccakWitness<F>,
    pub tables: Vec<MVLookupTable<F, LookupTableIDs>>,
    pub multiplicities: Vec<Vec<u32>>,
    /* private fields */
}

This struct contains all that needs to be kept track of during the execution of the Keccak step interpreter

Fields
witness: KeccakWitness<F>

The full state of the Keccak gate (witness)

tables: Vec<MVLookupTable<F, LookupTableIDs>>

The fixed tables used in the Keccak gate

multiplicities: Vec<Vec<u32>>

The multiplicities of each lookup entry. Should not be cleared between steps.