Type Alias kimchi_optimism::mips::folding::MIPSFoldingInstance

pub type MIPSFoldingInstance = FoldingInstance<MIPS_COLUMNS>;

Aliased Type

struct MIPSFoldingInstance {
    pub commitments: [GroupAffine<Parameters>; 166],
    pub challenges: [Fp256<FrParameters>; 3],
    pub alphas: Alphas,
}

Fields
commitments: [GroupAffine<Parameters>; 166]

Commitments to the witness columns, including the dynamic selectors

challenges: [Fp256<FrParameters>; 3]

Challenges for the proof. We do use 3 challenges:

β as the evaluation point for the logup argument

j: the joint combiner for vector lookups

γ (set to 0 for now)

alphas: Alphas

Reuses the Alphas defined in the example of folding