Trait kimchi_optimism::mips::interpreter::InterpreterEnv

pub trait InterpreterEnv {
    type Position;
    type Variable: Clone + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Mul<Self::Variable, Output = Self::Variable> + Debug;

    // Required methods
    fn alloc_scratch(&mut self) -> Self::Position;
    fn add_lookup(&mut self, lookup: Lookup<Self::Variable>);
    fn instruction_counter(&self) -> Self::Variable;
    unsafe fn fetch_register(
        &mut self,
        idx: &Self::Variable,
        output: Self::Position
    ) -> Self::Variable;
    unsafe fn push_register(
        &mut self,
        idx: &Self::Variable,
        value: Self::Variable
    );
    unsafe fn fetch_register_access(
        &mut self,
        idx: &Self::Variable,
        output: Self::Position
    ) -> Self::Variable;
    unsafe fn push_register_access(
        &mut self,
        idx: &Self::Variable,
        value: Self::Variable
    );
    unsafe fn fetch_memory(
        &mut self,
        addr: &Self::Variable,
        output: Self::Position
    ) -> Self::Variable;
    unsafe fn push_memory(
        &mut self,
        addr: &Self::Variable,
        value: Self::Variable
    );
    unsafe fn fetch_memory_access(
        &mut self,
        addr: &Self::Variable,
        output: Self::Position
    ) -> Self::Variable;
    unsafe fn push_memory_access(
        &mut self,
        addr: &Self::Variable,
        value: Self::Variable
    );
    fn set_instruction_pointer(&mut self, ip: Self::Variable);
    fn get_instruction_pointer(&self) -> Self::Variable;
    fn get_instruction_part(&self, part: InstructionPart) -> Self::Variable;
    fn constant(x: u32) -> Self::Variable;
    fn set_halted(&mut self, flag: Self::Variable);

    // Provided methods
    unsafe fn access_register(
        &mut self,
        idx: &Self::Variable,
        old_value: &Self::Variable,
        new_value: &Self::Variable
    ) { ... }
    fn read_register(&mut self, idx: &Self::Variable) -> Self::Variable { ... }
    fn write_register(
        &mut self,
        idx: &Self::Variable,
        new_value: Self::Variable
    ) { ... }
    unsafe fn access_memory(
        &mut self,
        addr: &Self::Variable,
        old_value: &Self::Variable,
        new_value: &Self::Variable
    ) { ... }
    fn read_memory(&mut self, addr: &Self::Variable) -> Self::Variable { ... }
    fn write_memory(&mut self, addr: &Self::Variable, new_value: Self::Variable) { ... }
    fn range_check64(&mut self, _value: &Self::Variable) { ... }
    fn get_immediate(&self) -> Self::Variable { ... }

}

Required Associated Types

type Position

type Variable: Clone + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Mul<Self::Variable, Output = Self::Variable> + Debug

Required Methods

fn alloc_scratch(&mut self) -> Self::Position

fn add_lookup(&mut self, lookup: Lookup<Self::Variable>)

fn instruction_counter(&self) -> Self::Variable

unsafe fn fetch_register( &mut self, idx: &Self::Variable, output: Self::Position ) -> Self::Variable

Fetch the value of the general purpose register with index idx and store it in local position output.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this operation.

unsafe fn push_register(&mut self, idx: &Self::Variable, value: Self::Variable)

Set the general purpose register with index idx to value.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this operation.

unsafe fn fetch_register_access( &mut self, idx: &Self::Variable, output: Self::Position ) -> Self::Variable

Fetch the last ‘access index’ for the general purpose register with index idx, and store it in local position output.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this operation.

unsafe fn push_register_access( &mut self, idx: &Self::Variable, value: Self::Variable )

Set the last ‘access index’ for the general purpose register with index idx to value.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this operation.

unsafe fn fetch_memory( &mut self, addr: &Self::Variable, output: Self::Position ) -> Self::Variable

Fetch the memory value at address addr and store it in local position output.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this memory operation.

unsafe fn push_memory(&mut self, addr: &Self::Variable, value: Self::Variable)

Set the memory value at address addr to value.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this memory operation.

unsafe fn fetch_memory_access( &mut self, addr: &Self::Variable, output: Self::Position ) -> Self::Variable

Fetch the last ‘access index’ that the memory at address addr was written at, and store it in local position output.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this memory operation.

unsafe fn push_memory_access( &mut self, addr: &Self::Variable, value: Self::Variable )

Set the last ‘access index’ for the memory at address addr to value.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this memory operation.

fn set_instruction_pointer(&mut self, ip: Self::Variable)

fn get_instruction_pointer(&self) -> Self::Variable

fn get_instruction_part(&self, part: InstructionPart) -> Self::Variable

fn constant(x: u32) -> Self::Variable

fn set_halted(&mut self, flag: Self::Variable)

Provided Methods

unsafe fn access_register( &mut self, idx: &Self::Variable, old_value: &Self::Variable, new_value: &Self::Variable )

Access the general purpose register with index idx, adding constraints asserting that the old value was old_value and that the new value will be new_value.

Safety

Callers of this function must manually update the registers if required, this function will only update the access counter.

fn read_register(&mut self, idx: &Self::Variable) -> Self::Variable

fn write_register(&mut self, idx: &Self::Variable, new_value: Self::Variable)

unsafe fn access_memory( &mut self, addr: &Self::Variable, old_value: &Self::Variable, new_value: &Self::Variable )

Access the memory address addr, adding constraints asserting that the old value was old_value and that the new value will be new_value.

Safety

Callers of this function must manually update the memory if required, this function will only update the access counter.

fn read_memory(&mut self, addr: &Self::Variable) -> Self::Variable

fn write_memory(&mut self, addr: &Self::Variable, new_value: Self::Variable)

fn range_check64(&mut self, _value: &Self::Variable)

fn get_immediate(&self) -> Self::Variable