Trait kimchi_optimism::trace::Tracer

pub trait Tracer<const COLUMNS: usize, SELECTOR, F: Zero, Env> {
    // Required methods
    fn new(domain_size: usize, env: &mut Env) -> Self;
    fn push_row(&mut self, opcode: SELECTOR, row: &[F; COLUMNS]);
    fn pad_with_row(&mut self, opcode: SELECTOR, row: &[F; COLUMNS]) -> usize;
    fn pad_with_zeros(&mut self, opcode: SELECTOR) -> usize;
    fn pad_dummy(&mut self, opcode: SELECTOR) -> usize;
    fn pad_witnesses(&mut self);
}

Tracer builds traces for some program executions. The constant type COLUMNS is defined as the maximum number of columns/“registers” the trace can use per row. The type SELECTOR encodes the information of the kind of information the trace encodes. Examples:

For Keccak, Step encodes the row being performed at a time: round, squeeze, padding, etc…

For MIPS, Instruction encodes the CPU instruction being executed: add, sub, load, store, etc… The type parameter F is the type the data points in the trace are encoded into. It can be a field or a native type (u64).

Required Methods

fn new(domain_size: usize, env: &mut Env) -> Self

Create a new circuit

fn push_row(&mut self, opcode: SELECTOR, row: &[F; COLUMNS])

Add a witness row to the circuit

fn pad_with_row(&mut self, opcode: SELECTOR, row: &[F; COLUMNS]) -> usize

Pad the rows of one opcode with the given row until reaching the domain size if needed. Returns the number of rows that were added.

fn pad_with_zeros(&mut self, opcode: SELECTOR) -> usize

Pads the rows of one opcode with zero rows until reaching the domain size if needed. Returns the number of rows that were added.

fn pad_dummy(&mut self, opcode: SELECTOR) -> usize

Pad the rows of one opcode with the first row until reaching the domain size if needed. It only tries to pad witnesses which are non empty. Returns the number of rows that were added.

fn pad_witnesses(&mut self)

Pads the rows of the witnesses until reaching the domain size using the first row repeatedly.

Object Safety

This trait is not object safe.