Trait kimchi_optimism::CircuitTrait

pub trait CircuitTrait<const N: usize, SELECTOR, F, Env> {
    // Required methods
    fn new(domain_size: usize, env: &mut Env) -> Self;
    fn push_row(&mut self, step: SELECTOR, row: &[F; N]);
    fn pad(&mut self, step: SELECTOR) -> bool;
    fn pad_witnesses(&mut self);
    fn reset(&mut self, step: SELECTOR);
}

Required Methods

fn new(domain_size: usize, env: &mut Env) -> Self

Create a new circuit

fn push_row(&mut self, step: SELECTOR, row: &[F; N])

Add a witness row to the circuit

fn pad(&mut self, step: SELECTOR) -> bool

Pads the rows of one selector until reaching the domain size if needed. Returns true if padding was performed, false otherwise.

fn pad_witnesses(&mut self)

Pads the rows of the witnesses until reaching the domain size

fn reset(&mut self, step: SELECTOR)

Resets the witness after folding

Object Safety

This trait is not object safe.