Type Alias kimchi_optimism::keccak::column::KeccakWitness

pub type KeccakWitness<T> = Witness<ZKVM_KECCAK_REL, T>;

The witness columns used by the Keccak circuit. The Keccak circuit is split into two main modes: Sponge and Round. The columns are shared between the Sponge and Round steps. The hash and step indices are shared between both modes. The row is split into the following entries:

mode_flags: what kind of mode is running: round, root, absorb, pad, rootpad, squeeze. Only 1 of them can be active.

hash_index: Which hash this is inside the circuit

block_index: Which block this is inside the hash

step_index: Which step this is inside the hash

curr: Contains 1965 witnesses used in the current step including Input

next: Contains the 100 Output witnesses

round_flags: contain 5 elements with information about the current round step

pad_flags: PadLength, TwoToPad, PadBytesFlags, PadSuffix

Keccak Witness Columns: KeccakWitness.cols

| 0..=5 | 6 | 7 | 8 | 9..1973 | 1974..2073 | 2074..2078 |

0..=5 -> mode_flags 6 -> hash_index 7 -> block_index 8 -> step_index 9..=1973 -> curr 9 1973 <––––––––––––––––if_round<–––––––––––––––––> <———––if_sponge–––––––> 9 808 -> SPONGE: | -> ROUND: -> 9..=108: Input == SpongeOldState | -> 9..=108: Input == ThetaStateA -> 109..=208: SpongeNewState | -> 109..=188: ThetaShiftsC : 176..=207 -> SpongeZeros | -> 189..=208: ThetaDenseC -> 209..=408: SpongeBytes | -> 209..=213: ThetaQuotientC -> 409..=808: SpongeShifts | -> 214..=233: ThetaRemainderC | -> 234..=253: ThetaDenseRotC | -> 254..=273: ThetaExpandRotC | -> 274..=673: PiRhoShiftsE | -> 674..=773: PiRhoDenseE | -> 774..=873: PiRhoQuotientE | -> 874..=973: PiRhoRemainderE | -> 974..=1073: PiRhoDenseRotE | -> 1074..=1173: PiRhoExpandRotE | -> 1174..=1573: ChiShiftsB | -> 1574..=1973: ChiShiftsSum 1974..=2073 -> next -> 1974..=2073: Output (if Round, then IotaStateG, if Sponge then SpongeXorState)

2074..=2078 -> round_flags -> 2074: RoundNumber -> 2075..=2078: RoundConstants

809..=951 -> pad_flags -> 809: PadLength -> 810: TwoToPad -> 811..=815: PadSuffix -> 816..=951: PadBytesFlags

Aliased Type

struct KeccakWitness<T> {
    pub cols: Box<[T; 2073]>,
}

Fields
cols: Box<[T; 2073]>

A witness row is represented by an array of N witness columns When T is a vector, then the witness describes the rows of the circuit.