Struct kimchi_optimism::keccak::environment::KeccakEnv

pub struct KeccakEnv<Fp> {
    pub keccak_witness: KeccakWitness<Fp>,
    pub keccak_step: Option<KeccakStep>,
    /* private fields */
}

This struct contains all that needs to be kept track of during the execution of the Keccak step interpreter

Fields
keccak_witness: KeccakWitness<Fp>

The full state of the Keccak gate (witness)

keccak_step: Option<KeccakStep>

What step of the hash is being executed (or None, if just ended)

Implementations

impl<Fp: Field> KeccakEnv<Fp>

pub fn new(hash_idx: u64, preimage: &[u8]) -> Self

Starts a new Keccak environment for a given hash index and bytestring of preimage data

pub fn write_column(&mut self, column: KeccakColumn, value: u64)

Writes an integer value to a column of the Keccak witness

pub fn write_column_field(&mut self, column: KeccakColumn, value: Fp)

Writes a field value to a column of the Keccak witness

pub fn null_state(&mut self)

Nullifies the KeccakWitness of the environment by resetting it to default values

pub fn update_step(&mut self)

This function updates the next step of the environment depending on the current step