Trait kimchi_optimism::keccak::interpreter::KeccakInterpreter

pub trait KeccakInterpreter<F: One + Debug + Zero> {
    type Variable: Mul<Self::Variable, Output = Self::Variable> + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Clone + Debug + One + Zero;

    // Required methods
    fn constant(x: u64) -> Self::Variable;
    fn constant_field(x: F) -> Self::Variable;
    fn two_pow(x: u64) -> Self::Variable;
    fn variable(&self, column: KeccakColumn) -> Self::Variable;
    fn constrain(&mut self, x: Self::Variable);
    fn add_lookup(&mut self, lookup: RAMLookup<Self::Variable, LookupTableIDs>);
    fn lookup_syscall_preimage(&mut self);
    fn lookup_syscall_hash(&mut self);
    fn lookup_steps(&mut self);
    fn lookup_rc16(&mut self, flag: Self::Variable, value: Self::Variable);
    fn lookup_reset(
        &mut self,
        flag: Self::Variable,
        dense: Self::Variable,
        sparse: Self::Variable
    );
    fn lookup_sparse(&mut self, flag: Self::Variable, value: Self::Variable);
    fn lookup_byte(&mut self, flag: Self::Variable, value: Self::Variable);
    fn lookup_pad(&mut self, flag: Self::Variable, value: Vec<Self::Variable>);
    fn lookup_round_constants(
        &mut self,
        flag: Self::Variable,
        value: Vec<Self::Variable>
    );

    // Provided methods
    fn is_boolean(x: Self::Variable) -> Self::Variable { ... }
    fn not(x: Self::Variable) -> Self::Variable { ... }
    fn is_one(x: Self::Variable) -> Self::Variable { ... }
    fn is_nonzero(x: Self::Variable, x_inv: Self::Variable) -> Self::Variable { ... }
    fn xor(x: Self::Variable, y: Self::Variable) -> Self::Variable { ... }
    fn or(x: Self::Variable, y: Self::Variable) -> Self::Variable { ... }
    fn either_zero(x: Self::Variable, y: Self::Variable) -> Self::Variable { ... }
    fn zero() -> Self::Variable { ... }
    fn one() -> Self::Variable { ... }
    fn two() -> Self::Variable { ... }
    fn constraints(&mut self) { ... }
    fn constrain_flags(&mut self) { ... }
    fn constrain_booleanity(&mut self) { ... }
    fn constrain_mutex(&mut self) { ... }
    fn constrain_sponge(&mut self) { ... }
    fn constrain_absorb(&mut self) { ... }
    fn constrain_padding(&mut self) { ... }
    fn constrain_squeeze(&mut self) { ... }
    fn constrain_round(&mut self) { ... }
    fn constrain_theta(&mut self) -> Vec<Vec<Vec<Self::Variable>>> { ... }
    fn constrain_pirho(
        &mut self,
        state_e: Vec<Vec<Vec<Self::Variable>>>
    ) -> Vec<Vec<Vec<Self::Variable>>> { ... }
    fn constrain_chi(
        &mut self,
        state_b: Vec<Vec<Vec<Self::Variable>>>
    ) -> Vec<Vec<Vec<Self::Variable>>> { ... }
    fn constrain_iota(&mut self, state_f: Vec<Vec<Vec<Self::Variable>>>) { ... }
    fn lookups(&mut self) { ... }
    fn lookups_sponge(&mut self) { ... }
    fn lookups_round_theta(&mut self) { ... }
    fn lookups_round_pirho(&mut self) { ... }
    fn lookups_round_chi(&mut self) { ... }
    fn lookups_round_iota(&mut self) { ... }
    fn from_shifts(
        shifts: &[Self::Variable],
        i: Option<usize>,
        y: Option<usize>,
        x: Option<usize>,
        q: Option<usize>
    ) -> Self::Variable { ... }
    fn from_quarters(
        quarters: &[Self::Variable],
        y: Option<usize>,
        x: usize
    ) -> Self::Variable { ... }
    fn is_sponge(&self) -> Self::Variable { ... }
    fn is_absorb(&self) -> Self::Variable { ... }
    fn is_squeeze(&self) -> Self::Variable { ... }
    fn is_root(&self) -> Self::Variable { ... }
    fn is_pad(&self) -> Self::Variable { ... }
    fn is_round(&self) -> Self::Variable { ... }
    fn round(&self) -> Self::Variable { ... }
    fn pad_length(&self) -> Self::Variable { ... }
    fn two_to_pad(&self) -> Self::Variable { ... }
    fn in_padding(&self, idx: usize) -> Self::Variable { ... }
    fn pad_suffix(&self, idx: usize) -> Self::Variable { ... }
    fn bytes_block(&self, idx: usize) -> Vec<Self::Variable> { ... }
    fn pad_bytes_flags(&self) -> [Self::Variable; 136] { ... }
    fn flags_block(&self, idx: usize) -> Vec<Self::Variable> { ... }
    fn block_in_padding(&self, idx: usize) -> Self::Variable { ... }
    fn round_constants(&self) -> [Self::Variable; 4] { ... }
    fn old_state(&self, idx: usize) -> Self::Variable { ... }
    fn new_state(&self, idx: usize) -> Self::Variable { ... }
    fn xor_state(&self, idx: usize) -> Self::Variable { ... }
    fn sponge_zeros(&self) -> [Self::Variable; 32] { ... }
    fn vec_sponge_shifts(&self) -> [Self::Variable; 400] { ... }
    fn sponge_shifts(&self, idx: usize) -> Self::Variable { ... }
    fn sponge_bytes(&self) -> [Self::Variable; 200] { ... }
    fn sponge_byte(&self, idx: usize) -> Self::Variable { ... }
    fn state_a(&self, y: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_shifts_c(&self) -> [Self::Variable; 80] { ... }
    fn shifts_c(&self, i: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_dense_c(&self) -> [Self::Variable; 20] { ... }
    fn dense_c(&self, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_quotient_c(&self) -> [Self::Variable; 5] { ... }
    fn quotient_c(&self, x: usize) -> Self::Variable { ... }
    fn vec_remainder_c(&self) -> [Self::Variable; 20] { ... }
    fn remainder_c(&self, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_dense_rot_c(&self) -> [Self::Variable; 20] { ... }
    fn dense_rot_c(&self, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_expand_rot_c(&self) -> [Self::Variable; 20] { ... }
    fn expand_rot_c(&self, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_shifts_e(&self) -> [Self::Variable; 400] { ... }
    fn shifts_e(&self, i: usize, y: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_dense_e(&self) -> [Self::Variable; 100] { ... }
    fn dense_e(&self, y: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_quotient_e(&self) -> [Self::Variable; 100] { ... }
    fn quotient_e(&self, y: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_remainder_e(&self) -> [Self::Variable; 100] { ... }
    fn remainder_e(&self, y: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_dense_rot_e(&self) -> [Self::Variable; 100] { ... }
    fn dense_rot_e(&self, y: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_expand_rot_e(&self) -> [Self::Variable; 100] { ... }
    fn expand_rot_e(&self, y: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_shifts_b(&self) -> [Self::Variable; 400] { ... }
    fn shifts_b(&self, i: usize, y: usize, x: usize, q: usize) -> Self::Variable { ... }
    fn vec_shifts_sum(&self) -> [Self::Variable; 400] { ... }
    fn shifts_sum(
        &self,
        i: usize,
        y: usize,
        x: usize,
        q: usize
    ) -> Self::Variable { ... }
    fn state_g(&self, idx: usize) -> Self::Variable { ... }
    fn hash_index(&self) -> Self::Variable { ... }
    fn block_index(&self) -> Self::Variable { ... }
    fn step_index(&self) -> Self::Variable { ... }
    fn input(&self) -> [Self::Variable; 100] { ... }
    fn input_of_step(&self) -> Vec<Self::Variable> { ... }
    fn output(&self) -> [Self::Variable; 100] { ... }
    fn output_of_step(&self) -> Vec<Self::Variable> { ... }

}

This trait includes functionalities needed to obtain the variables of the Keccak circuit needed for constraints and witness

Required Associated Types

type Variable: Mul<Self::Variable, Output = Self::Variable> + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Clone + Debug + One + Zero

Required Methods

fn constant(x: u64) -> Self::Variable

Creates a variable from a constant integer

fn constant_field(x: F) -> Self::Variable

Creates a variable from a constant field element

fn two_pow(x: u64) -> Self::Variable

Returns a variable representing the value 2^x

fn variable(&self, column: KeccakColumn) -> Self::Variable

Returns the variable corresponding to a given column alias.

fn constrain(&mut self, x: Self::Variable)

Adds one constraint to the environment.

fn add_lookup(&mut self, lookup: RAMLookup<Self::Variable, LookupTableIDs>)

Adds a given Lookup to the environment

fn lookup_syscall_preimage(&mut self)

Reads Lookups containing the 136 bytes of the block of the preimage

fn lookup_syscall_hash(&mut self)

Writes a Lookup containing the 31byte output of the hash (excludes the MSB)

fn lookup_steps(&mut self)

Reads a Lookup containing the input of a step and writes a Lookup containing the output of the next step

fn lookup_rc16(&mut self, flag: Self::Variable, value: Self::Variable)

Adds a lookup to the RangeCheck16 table

fn lookup_reset( &mut self, flag: Self::Variable, dense: Self::Variable, sparse: Self::Variable )

Adds a lookup to the Reset table

fn lookup_sparse(&mut self, flag: Self::Variable, value: Self::Variable)

Adds a lookup to the Shift table

fn lookup_byte(&mut self, flag: Self::Variable, value: Self::Variable)

Adds a lookup to the Byte table

fn lookup_pad(&mut self, flag: Self::Variable, value: Vec<Self::Variable>)

Adds a lookup to the Pad table

fn lookup_round_constants( &mut self, flag: Self::Variable, value: Vec<Self::Variable> )

Adds a lookup to the RoundConstants table

Provided Methods

fn is_boolean(x: Self::Variable) -> Self::Variable

Degree-2 variable encoding whether the input is a boolean value (0 = yes)

fn not(x: Self::Variable) -> Self::Variable

Degree-1 variable encoding the negation of the input Note: it only works as expected if the input is a boolean value

fn is_one(x: Self::Variable) -> Self::Variable

Degree-1 variable encoding whether the input is the value one (0 = yes)

fn is_nonzero(x: Self::Variable, x_inv: Self::Variable) -> Self::Variable

Degree-2 variable encoding whether the first input is nonzero (0 = yes). It requires the second input to be the multiplicative inverse of the first. Note: if the first input is zero, there is no multiplicative inverse.

fn xor(x: Self::Variable, y: Self::Variable) -> Self::Variable

Degree-1 variable encoding the XOR of two variables which should be boolean (1 = true)

fn or(x: Self::Variable, y: Self::Variable) -> Self::Variable

Degree-1 variable encoding the OR of two variables, which should be boolean (1 = true)

fn either_zero(x: Self::Variable, y: Self::Variable) -> Self::Variable

Degree-2 variable encoding whether at least one of the two inputs is zero (0 = yes)

fn zero() -> Self::Variable

Returns a variable representing the value zero

fn one() -> Self::Variable

Returns a variable representing the value one

fn two() -> Self::Variable

Returns a variable representing the value two

fn constraints(&mut self)

Adds all 887 constraints/checks to the environment:

143 constraints of degree 1

739 constraints of degree 2

5 constraints of degree 5

fn constrain_flags(&mut self)

Constrains 144 checks of correctness of mode flags

143 constraints of degree 1

1 constraint of degree 2 Of which:

142 constraints are sponge-only

1 constraint is sponge+round related

fn constrain_booleanity(&mut self)

Constrains 139 checks of booleanity for some mode flags. These involve sponge-only related variables.

fn constrain_mutex(&mut self)

Constrains 5 checks of mutual exclusivity between some mode flags.

4 involve sponge-only related variables

1 involves sponge+round variables

fn constrain_sponge(&mut self)

Constrains 354 checks of sponge steps

fn constrain_absorb(&mut self)

Constrains 332 checks of absorb sponges

fn constrain_padding(&mut self)

Constrains 6 checks of padding absorb sponges

fn constrain_squeeze(&mut self)

Constrains 16 checks of squeeze sponges

fn constrain_round(&mut self)

Constrains 389 checks of round steps

384 constraints of degree 2

5 constraints of degree 3

fn constrain_theta(&mut self) -> Vec<Vec<Vec<Self::Variable>>>

Constrains 35 checks of the theta algorithm in round steps

30 constraints of degree 2

5 constraints of degree 3

fn constrain_pirho( &mut self, state_e: Vec<Vec<Vec<Self::Variable>>> ) -> Vec<Vec<Vec<Self::Variable>>>

Constrains 150 checks (of degree 2) of the pirho algorithm in round steps

fn constrain_chi( &mut self, state_b: Vec<Vec<Vec<Self::Variable>>> ) -> Vec<Vec<Vec<Self::Variable>>>

Constrains 200 checks (of degree 2) of the chi algorithm in round steps

fn constrain_iota(&mut self, state_f: Vec<Vec<Vec<Self::Variable>>>)

Constrains 4 checks (of degree 2) of the iota algorithm in round steps

fn lookups(&mut self)

Adds all 2481 lookups to the Keccak constraints environment:

2342 lookups for the step row

2 lookups for the inter-step channel

136 lookups for the syscall channel (preimage bytes)

1 lookups for the syscall channel (hash)

fn lookups_sponge(&mut self)

Adds the 601 lookups required for the sponge

fn lookups_round_theta(&mut self)

Adds the 140 lookups required for Theta in the round

fn lookups_round_pirho(&mut self)

Adds the 800 lookups required for PiRho in the round

fn lookups_round_chi(&mut self)

Adds the 800 lookups required for Chi in the round

fn lookups_round_iota(&mut self)

Adds the 1 lookup required for Iota in the round

fn from_shifts( shifts: &[Self::Variable], i: Option<usize>, y: Option<usize>, x: Option<usize>, q: Option<usize> ) -> Self::Variable

COLUMN OPERATIONS /// This function returns the composed sparse variable from shifts of any correct length:

When the length is 400, two index configurations are possible:

If i is Some, then this sole index could range between [0..400)

If i is None, then y, x and q must be Some and

y must range between [0..5)

x must range between [0..5)

q must range between [0..4)

When the length is 80, both i and y should be None, and x and q must be Some with:

x must range between [0..5)

q must range between [0..4)

fn from_quarters( quarters: &[Self::Variable], y: Option<usize>, x: usize ) -> Self::Variable

This function returns the composed variable from dense quarters of any correct length:

When y is Some, then the length must be 100 and:

y must range between [0..5)

x must range between [0..5)

When y is None, then the length must be 20 and:

x must range between [0..5)

fn is_sponge(&self) -> Self::Variable

Returns a variable that encodes whether the current step is a sponge (1 = yes)

fn is_absorb(&self) -> Self::Variable

Returns a variable that encodes whether the current step is an absorb sponge (1 = yes)

fn is_squeeze(&self) -> Self::Variable

Returns a variable that encodes whether the current step is a squeeze sponge (1 = yes)

fn is_root(&self) -> Self::Variable

Returns a variable that encodes whether the current step is the first absorb sponge (1 = yes)

fn is_pad(&self) -> Self::Variable

Returns a degree-2 variable that encodes whether the current step is the last absorb sponge (1 = yes)

fn is_round(&self) -> Self::Variable

Returns a variable that encodes whether the current step is a permutation round (1 = yes)

fn round(&self) -> Self::Variable

Returns a variable that encodes the current round number [0..24)

fn pad_length(&self) -> Self::Variable

Returns a variable that encodes the bytelength of the padding if any [0..136)

fn two_to_pad(&self) -> Self::Variable

Returns a variable that encodes the value 2^pad_length

fn in_padding(&self, idx: usize) -> Self::Variable

Returns a variable that encodes whether the idx-th byte of the new block is involved in the padding (1 = yes)

fn pad_suffix(&self, idx: usize) -> Self::Variable

Returns a variable that encodes the idx-th chunk of the padding suffix

if idx = 0, then the length is 12 bytes at most

if idx = [1..5), then the length is 31 bytes at most

fn bytes_block(&self, idx: usize) -> Vec<Self::Variable>

Returns a variable that encodes the idx-th block of bytes of the new block by composing the bytes variables, with idx in [0..5)

fn pad_bytes_flags(&self) -> [Self::Variable; 136]

Returns the 136 flags indicating which bytes of the new block are involved in the padding, as variables

fn flags_block(&self, idx: usize) -> Vec<Self::Variable>

Returns a vector of pad bytes flags as variables, with idx in [0..5)

if idx = 0, then the length of the block is at most 12

if idx = [1..5), then the length of the block is at most 31

fn block_in_padding(&self, idx: usize) -> Self::Variable

This function returns a variable that is computed as the accumulated value of the operation byte * flag * 2^8 for each byte block and flag block of the new block. This function will be used in constraints to determine whether the padding is located at the end of the preimage data, as consecutive bits that are involved in the padding.

fn round_constants(&self) -> [Self::Variable; 4]

Returns the 4 expanded quarters that encode the round constant, as variables

fn old_state(&self, idx: usize) -> Self::Variable

Returns the idx-th old state expanded quarter, as a variable

fn new_state(&self, idx: usize) -> Self::Variable

Returns the idx-th new state expanded quarter, as a variable

fn xor_state(&self, idx: usize) -> Self::Variable

Returns the output of an absorb sponge, which is the XOR of the old state and the new state

fn sponge_zeros(&self) -> [Self::Variable; 32]

Returns the last 32 terms that are added to the new block in an absorb sponge, as variables which should be zeros

fn vec_sponge_shifts(&self) -> [Self::Variable; 400]

Returns the 400 terms that compose the shifts of the sponge, as variables

fn sponge_shifts(&self, idx: usize) -> Self::Variable

Returns the idx-th term of the shifts of the sponge, as a variable

fn sponge_bytes(&self) -> [Self::Variable; 200]

Returns the 200 bytes of the sponge, as variables

fn sponge_byte(&self, idx: usize) -> Self::Variable

Returns the idx-th byte of the sponge, as a variable

fn state_a(&self, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (y,x,q)-th input of the theta algorithm, as a variable

fn vec_shifts_c(&self) -> [Self::Variable; 80]

Returns the 80 variables corresponding to ThetaShiftsC

fn shifts_c(&self, i: usize, x: usize, q: usize) -> Self::Variable

Returns the (i,x,q)-th variable of ThetaShiftsC

fn vec_dense_c(&self) -> [Self::Variable; 20]

Returns the 20 variables corresponding to ThetaDenseC

fn dense_c(&self, x: usize, q: usize) -> Self::Variable

Returns the (x,q)-th term of ThetaDenseC, as a variable

fn vec_quotient_c(&self) -> [Self::Variable; 5]

Returns the 5 variables corresponding to ThetaQuotientC

fn quotient_c(&self, x: usize) -> Self::Variable

Returns the (x)-th term of ThetaQuotientC, as a variable

fn vec_remainder_c(&self) -> [Self::Variable; 20]

Returns the 20 variables corresponding to ThetaRemainderC

fn remainder_c(&self, x: usize, q: usize) -> Self::Variable

Returns the (x,q)-th variable of ThetaRemainderC

fn vec_dense_rot_c(&self) -> [Self::Variable; 20]

Returns the 20 variables corresponding to ThetaDenseRotC

fn dense_rot_c(&self, x: usize, q: usize) -> Self::Variable

Returns the (x,q)-th variable of ThetaDenseRotC

fn vec_expand_rot_c(&self) -> [Self::Variable; 20]

Returns the 20 variables corresponding to ThetaExpandRotC

fn expand_rot_c(&self, x: usize, q: usize) -> Self::Variable

Returns the (x,q)-th variable of ThetaExpandRotC

fn vec_shifts_e(&self) -> [Self::Variable; 400]

Returns the 400 variables corresponding to PiRhoShiftsE

fn shifts_e(&self, i: usize, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (i,y,x,q)-th variable of PiRhoShiftsE

fn vec_dense_e(&self) -> [Self::Variable; 100]

Returns the 100 variables corresponding to PiRhoDenseE

fn dense_e(&self, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (y,x,q)-th variable of PiRhoDenseE

fn vec_quotient_e(&self) -> [Self::Variable; 100]

Returns the 100 variables corresponding to PiRhoQuotientE

fn quotient_e(&self, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (y,x,q)-th variable of PiRhoQuotientE

fn vec_remainder_e(&self) -> [Self::Variable; 100]

Returns the 100 variables corresponding to PiRhoRemainderE

fn remainder_e(&self, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (y,x,q)-th variable of PiRhoRemainderE

fn vec_dense_rot_e(&self) -> [Self::Variable; 100]

Returns the 100 variables corresponding to PiRhoDenseRotE

fn dense_rot_e(&self, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (y,x,q)-th variable of PiRhoDenseRotE

fn vec_expand_rot_e(&self) -> [Self::Variable; 100]

Returns the 100 variables corresponding to PiRhoExpandRotE

fn expand_rot_e(&self, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (y,x,q)-th variable of PiRhoExpandRotE

fn vec_shifts_b(&self) -> [Self::Variable; 400]

Returns the 400 variables corresponding to ChiShiftsB

fn shifts_b(&self, i: usize, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (i,y,x,q)-th variable of ChiShiftsB

fn vec_shifts_sum(&self) -> [Self::Variable; 400]

Returns the 400 variables corresponding to ChiShiftsSum

fn shifts_sum(&self, i: usize, y: usize, x: usize, q: usize) -> Self::Variable

Returns the (i,y,x,q)-th variable of ChiShiftsSum

fn state_g(&self, idx: usize) -> Self::Variable

Returns the idx-th output of a round step as a variable

fn hash_index(&self) -> Self::Variable

Returns the hash index as a variable

fn block_index(&self) -> Self::Variable

Returns the block index as a variable

fn step_index(&self) -> Self::Variable

Returns the step index as a variable

fn input(&self) -> [Self::Variable; 100]

Returns the 100 step input variables, which correspond to the:

State A when the current step is a permutation round

Old state when the current step is a non-root sponge

fn input_of_step(&self) -> Vec<Self::Variable>

Returns a slice of the input variables of the current step including the current hash index and step index

fn output(&self) -> [Self::Variable; 100]

Returns the 100 step output variables, which correspond to the:

State G when the current step is a permutation round

Xor state when the current step is an absorb sponge

fn output_of_step(&self) -> Vec<Self::Variable>

Returns a slice of the output variables of the current step (= input of next step) including the current hash index and step index

Object Safety

This trait is not object safe.