Trait kimchi_optimism::mips::interpreter::InterpreterEnv

pub trait InterpreterEnv {
    type Position;
    type Variable: Clone + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Mul<Self::Variable, Output = Self::Variable> + Debug;

    // Required methods
    fn alloc_scratch(&mut self) -> Self::Position;
    fn add_lookup(&mut self, lookup: Lookup<Self::Variable>);
    fn instruction_counter(&self) -> Self::Variable;
    unsafe fn fetch_register(
        &mut self,
        idx: &Self::Variable,
        output: Self::Position
    ) -> Self::Variable;
    unsafe fn push_register(
        &mut self,
        idx: &Self::Variable,
        value: Self::Variable
    );
    unsafe fn fetch_register_access(
        &mut self,
        idx: &Self::Variable,
        output: Self::Position
    ) -> Self::Variable;
    unsafe fn push_register_access(
        &mut self,
        idx: &Self::Variable,
        value: Self::Variable
    );
    unsafe fn fetch_memory(
        &mut self,
        addr: &Self::Variable,
        output: Self::Position
    ) -> Self::Variable;
    unsafe fn push_memory(
        &mut self,
        addr: &Self::Variable,
        value: Self::Variable
    );
    unsafe fn fetch_memory_access(
        &mut self,
        addr: &Self::Variable,
        output: Self::Position
    ) -> Self::Variable;
    unsafe fn push_memory_access(
        &mut self,
        addr: &Self::Variable,
        value: Self::Variable
    );
    fn constant(x: u32) -> Self::Variable;
    unsafe fn bitmask(
        &mut self,
        x: &Self::Variable,
        highest_bit: u32,
        lowest_bit: u32,
        position: Self::Position
    ) -> Self::Variable;
    unsafe fn shift_left(
        &mut self,
        x: &Self::Variable,
        by: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    unsafe fn shift_right(
        &mut self,
        x: &Self::Variable,
        by: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    unsafe fn shift_right_arithmetic(
        &mut self,
        x: &Self::Variable,
        by: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    unsafe fn test_zero(
        &mut self,
        x: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    unsafe fn test_less_than(
        &mut self,
        x: &Self::Variable,
        y: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    unsafe fn test_less_than_signed(
        &mut self,
        x: &Self::Variable,
        y: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    unsafe fn or_witness(
        &mut self,
        x: &Self::Variable,
        y: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    unsafe fn xor_witness(
        &mut self,
        x: &Self::Variable,
        y: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    fn copy(
        &mut self,
        x: &Self::Variable,
        position: Self::Position
    ) -> Self::Variable;
    fn set_halted(&mut self, flag: Self::Variable);

    // Provided methods
    unsafe fn access_register(
        &mut self,
        idx: &Self::Variable,
        old_value: &Self::Variable,
        new_value: &Self::Variable
    ) { ... }
    fn read_register(&mut self, idx: &Self::Variable) -> Self::Variable { ... }
    fn write_register(
        &mut self,
        idx: &Self::Variable,
        new_value: Self::Variable
    ) { ... }
    unsafe fn access_memory(
        &mut self,
        addr: &Self::Variable,
        old_value: &Self::Variable,
        new_value: &Self::Variable
    ) { ... }
    fn read_memory(&mut self, addr: &Self::Variable) -> Self::Variable { ... }
    fn write_memory(&mut self, addr: &Self::Variable, new_value: Self::Variable) { ... }
    fn range_check64(&mut self, _value: &Self::Variable) { ... }
    fn set_instruction_pointer(&mut self, ip: Self::Variable) { ... }
    fn get_instruction_pointer(&mut self) -> Self::Variable { ... }
    fn set_next_instruction_pointer(&mut self, ip: Self::Variable) { ... }
    fn get_next_instruction_pointer(&mut self) -> Self::Variable { ... }
    fn sign_extend(
        &mut self,
        x: &Self::Variable,
        bitlength: u32
    ) -> Self::Variable { ... }

}

Required Associated Types

type Position

type Variable: Clone + Add<Self::Variable, Output = Self::Variable> + Sub<Self::Variable, Output = Self::Variable> + Mul<Self::Variable, Output = Self::Variable> + Debug

Required Methods

fn alloc_scratch(&mut self) -> Self::Position

fn add_lookup(&mut self, lookup: Lookup<Self::Variable>)

fn instruction_counter(&self) -> Self::Variable

unsafe fn fetch_register( &mut self, idx: &Self::Variable, output: Self::Position ) -> Self::Variable

Fetch the value of the general purpose register with index idx and store it in local position output.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this operation.

unsafe fn push_register(&mut self, idx: &Self::Variable, value: Self::Variable)

Set the general purpose register with index idx to value.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this operation.

unsafe fn fetch_register_access( &mut self, idx: &Self::Variable, output: Self::Position ) -> Self::Variable

Fetch the last ‘access index’ for the general purpose register with index idx, and store it in local position output.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this operation.

unsafe fn push_register_access( &mut self, idx: &Self::Variable, value: Self::Variable )

Set the last ‘access index’ for the general purpose register with index idx to value.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this operation.

unsafe fn fetch_memory( &mut self, addr: &Self::Variable, output: Self::Position ) -> Self::Variable

Fetch the memory value at address addr and store it in local position output.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this memory operation.

unsafe fn push_memory(&mut self, addr: &Self::Variable, value: Self::Variable)

Set the memory value at address addr to value.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this memory operation.

unsafe fn fetch_memory_access( &mut self, addr: &Self::Variable, output: Self::Position ) -> Self::Variable

Fetch the last ‘access index’ that the memory at address addr was written at, and store it in local position output.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this memory operation.

unsafe fn push_memory_access( &mut self, addr: &Self::Variable, value: Self::Variable )

Set the last ‘access index’ for the memory at address addr to value.

Safety

No lookups or other constraints are added as part of this operation. The caller must manually add the lookups for this memory operation.

fn constant(x: u32) -> Self::Variable

unsafe fn bitmask( &mut self, x: &Self::Variable, highest_bit: u32, lowest_bit: u32, position: Self::Position ) -> Self::Variable

Extract the bits from the variable x between highest_bit and lowest_bit, and store the result in position. lowest_bit becomes the least-significant bit of the resulting value.

Safety

There are no constraints on the returned value; callers must assert the relationship with the source variable x and that the returned value fits in highest_bit - lowest_bit bits.

unsafe fn shift_left( &mut self, x: &Self::Variable, by: &Self::Variable, position: Self::Position ) -> Self::Variable

Return the result of shifting x by by, storing the result in position.

Safety

There are no constraints on the returned value; callers must assert the relationship with the source variable x and the shift amount by.

unsafe fn shift_right( &mut self, x: &Self::Variable, by: &Self::Variable, position: Self::Position ) -> Self::Variable

Return the result of shifting x by by, storing the result in position.

Safety

There are no constraints on the returned value; callers must assert the relationship with the source variable x and the shift amount by.

unsafe fn shift_right_arithmetic( &mut self, x: &Self::Variable, by: &Self::Variable, position: Self::Position ) -> Self::Variable

Return the result of shifting x by by, storing the result in position.

Safety

There are no constraints on the returned value; callers must assert the relationship with the source variable x and the shift amount by.

unsafe fn test_zero( &mut self, x: &Self::Variable, position: Self::Position ) -> Self::Variable

Returns 1 if x is 0, or 0 otherwise, storing the result in position.

Safety

There are no constraints on the returned value; callers must assert the relationship with x.

unsafe fn test_less_than( &mut self, x: &Self::Variable, y: &Self::Variable, position: Self::Position ) -> Self::Variable

Returns 1 if x < y as unsigned integers, or 0 otherwise, storing the result in position.

Safety

There are no constraints on the returned value; callers must assert that the value correctly represents the relationship between x and y

unsafe fn test_less_than_signed( &mut self, x: &Self::Variable, y: &Self::Variable, position: Self::Position ) -> Self::Variable

Returns 1 if x < y as signed integers, or 0 otherwise, storing the result in position.

Safety

There are no constraints on the returned value; callers must assert that the value correctly represents the relationship between x and y

unsafe fn or_witness( &mut self, x: &Self::Variable, y: &Self::Variable, position: Self::Position ) -> Self::Variable

Returns x or y, storing the result in position.

Safety

There are no constraints on the returned value; callers must manually add constraints to ensure that it is correctly constructed.

unsafe fn xor_witness( &mut self, x: &Self::Variable, y: &Self::Variable, position: Self::Position ) -> Self::Variable

Returns x xor y, storing the result in position.

Safety

There are no constraints on the returned value; callers must manually add constraints to ensure that it is correctly constructed.

fn copy( &mut self, x: &Self::Variable, position: Self::Position ) -> Self::Variable

fn set_halted(&mut self, flag: Self::Variable)

Provided Methods

unsafe fn access_register( &mut self, idx: &Self::Variable, old_value: &Self::Variable, new_value: &Self::Variable )

Access the general purpose register with index idx, adding constraints asserting that the old value was old_value and that the new value will be new_value.

Safety

Callers of this function must manually update the registers if required, this function will only update the access counter.

fn read_register(&mut self, idx: &Self::Variable) -> Self::Variable

fn write_register(&mut self, idx: &Self::Variable, new_value: Self::Variable)

unsafe fn access_memory( &mut self, addr: &Self::Variable, old_value: &Self::Variable, new_value: &Self::Variable )

Access the memory address addr, adding constraints asserting that the old value was old_value and that the new value will be new_value.

Safety

Callers of this function must manually update the memory if required, this function will only update the access counter.

fn read_memory(&mut self, addr: &Self::Variable) -> Self::Variable

fn write_memory(&mut self, addr: &Self::Variable, new_value: Self::Variable)

fn range_check64(&mut self, _value: &Self::Variable)

fn set_instruction_pointer(&mut self, ip: Self::Variable)

fn get_instruction_pointer(&mut self) -> Self::Variable

fn set_next_instruction_pointer(&mut self, ip: Self::Variable)

fn get_next_instruction_pointer(&mut self) -> Self::Variable

fn sign_extend(&mut self, x: &Self::Variable, bitlength: u32) -> Self::Variable